tuning of various parameters/added vibration animation to stunned parasites.
adjusting of hyper stock replenishment time as well as stun time, slightly.
will continue testing/adjusting to achieve the best possible balance of fun and challenge.

+++ final_project_B/Antibody.pde
@@ -7,7 +7,7 @@
 class Antibody {
 
   int maxEnergy = 255;
-  float moveEnergy = -2;
+  int moveEnergy = -2;
   int collideEnergy;
   int size;
   int energy;
